requirements.txt
flask
torch
torchvision
pillow
numpy
pandas
matplotlib
plotly
